refactor(vscode): lift a11y legend + overlay rendering out of behavior.ts
Slice of #767 — extracts the per-card accessibility overlay DOM
helpers (`buildA11yLegend`, `buildA11yOverlay`, `ensureHierarchyOverlay`,
`applyHierarchyOverlay`, plus the private `highlightA11yFinding`
hover handler) into a typed `a11yOverlay.ts` module.

`behavior.ts` keeps the per-preview caches (`cardA11yFindings` /
`cardA11yNodes`), the `applyA11yUpdate` orchestration, and the
`updateImage` rebuild path — those read `allPreviews` /
`filterToolbar` / `renderFocusInspector` and don't fit a pure-DOM
module.

Two minor cleanups in the lift:

- `buildA11yLegend(p)` becomes `buildA11yLegend(card, p)`. The hover
  callback now uses the card it was wired into instead of doing a
  cross-card `getElementById` lookup keyed off `previewId` — same
  effect, fewer DOM round-trips, and one less reason for the legend
  to know its preview's id beyond setting `data-preview-id` on the
  row.
- `parseBounds` lives privately in the new module. Same shape as the
  imperative copy; once #770 (which also lifts `parseBounds` into
  `cardData.ts`) lands a follow-up will dedupe.

`behavior.ts` shrinks from 3208 to 3057 lines (-151 LOC). No
behaviour change.

Refs #767.
